Move Slack commands to separate files
