+++ api/slack/commands/blacklist.py
@@ -1,0 +1,30 @@
+from api.slack import client
+from api.slack.utils import parse_user
+from blossom.strings import translation
+
+i18n = translation()
+
+
+def blacklist_cmd(channel: str, message: str) -> None:
+    """Blacklist a user based on a message from slack."""
+    parsed_message = message.split()
+    if len(parsed_message) == 1:
+        # they didn't give a username
+        msg = i18n["slack"]["errors"]["missing_username"]
+    elif len(parsed_message) == 2:
+        user, username = parse_user(parsed_message[1])
+        if user:
+            if user.blacklisted:
+                user.blacklisted = False
+                user.save()
+                msg = i18n["slack"]["blacklist"]["success_undo"].format(username)
+            else:
+                user.blacklisted = True
+                user.save()
+                msg = i18n["slack"]["blacklist"]["success"].format(username)
+        else:
+            msg = i18n["slack"]["errors"]["unknown_username"].format(username=username)
+    else:
+        msg = i18n["slack"]["errors"]["too_many_params"]
+
+    client.chat_postMessage(channel=channel, text=msg)
